Better error handling in wrapper script

+++ dt.sh
@@ -1,8 +1,13 @@
 #!/bin/env bash
 
+NEWDIRFILE=.newdir.dt
+
 function dt() {
     ./dt
-    prevfile=$(realpath .newdir.dt)
-    cd $(cat .newdir.dt)
+    if [[ ! -f $NEWDIRFILE ]]; then
+        return 0
+    fi
+    prevfile=$(realpath $NEWDIRFILE)
+    cd $(cat $NEWDIRFILE)
     rm "$prevfile"
 }
